Convert toolbelt handler to use typescript
